+++ transport/x/grpc/integration_test.go
@@ -36,11 +36,11 @@ import (
 	"go.uber.org/yarpc/internal/clientconfig"
 	"go.uber.org/yarpc/internal/examples/protobuf/example"
 	"go.uber.org/yarpc/internal/examples/protobuf/examplepb"
+	"go.uber.org/yarpc/internal/grpcctx"
 	"go.uber.org/yarpc/internal/testtime"
 	"go.uber.org/yarpc/yarpcerrors"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
 )
 
@@ -65,16 +65,6 @@ func TestGRPCBasic(t *testing.T) {
 		value, err := e.GetValueGRPC(context.Background(), "foo")
 		assert.NoError(t, err)
 		assert.Equal(t, "bar", value)
-	})
-}
-
-func TestYARPCMetadata(t *testing.T) {
-	t.Parallel()
-	var md metadata.MD
-	doWithTestEnv(t, nil, []InboundOption{withInboundUnaryInterceptor(newMetadataUnaryServerInterceptor(&md))}, nil, func(t *testing.T, e *testEnv) {
-		assert.NoError(t, e.SetValueYARPC(context.Background(), "foo", "bar"))
-		assert.Len(t, md["user-agent"], 1)
-		assert.True(t, strings.Contains(md["user-agent"][0], UserAgent))
 	})
 }
 
@@ -164,7 +154,7 @@ type testEnv struct {
 	Inbound             *Inbound
 	Outbound            *Outbound
 	ClientConn          *grpc.ClientConn
-	ContextWrapper      *ContextWrapper
+	ContextWrapper      *grpcctx.ContextWrapper
 	ClientConfig        transport.ClientConfig
 	Procedures          []transport.Procedure
 	KeyValueGRPCClient  examplepb.KeyValueClient
@@ -233,7 +223,7 @@ func newTestEnv(transportOptions []TransportOption, inboundOptions []InboundOpti
 	)
 	keyValueYARPCClient := examplepb.NewKeyValueYARPCClient(clientConfig)
 
-	contextWrapper := NewContextWrapper().
+	contextWrapper := grpcctx.NewContextWrapper().
 		WithCaller("example-client").
 		WithService("example").
 		WithEncoding(string(protobuf.Encoding))